+++ dreego-core/middleware_compress.go
@@ -1,0 +1,32 @@
+package core
+
+import (
+	"compress/gzip"
+	"net/http"
+	"strings"
+)
+
+func Compress() func(http.Handler) http.Handler {
+	return func(next http.Handler) http.Handler {
+		return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+			if !strings.Contains(r.Header.Get("Accept-Encoding"), "gzip") {
+				next.ServeHTTP(w, r)
+				return
+			}
+			w.Header().Set("Content-Encoding", "gzip")
+			gw := gzip.NewWriter(w)
+			defer gw.Close()
+			gwr := &gzipResponseWriter{ResponseWriter: w, Writer: gw}
+			next.ServeHTTP(gwr, r)
+		})
+	}
+}
+
+type gzipResponseWriter struct {
+	http.ResponseWriter
+	Writer *gzip.Writer
+}
+
+func (w *gzipResponseWriter) Write(b []byte) (int, error) {
+	return w.Writer.Write(b)
+}