+++ dreego-core/middleware_health.go
@@ -1,0 +1,31 @@
+package core
+
+import "net/http"
+
+var ready = true
+
+func SetReady(r bool) {
+	ready = r
+}
+
+func healthHandler() http.HandlerFunc {
+	return func(w http.ResponseWriter, _ *http.Request) {
+		w.Header().Set("Content-Type", "text/plain; charset=utf-8")
+		w.WriteHeader(http.StatusOK)
+		w.Write([]byte("ok"))
+	}
+}
+
+func readyHandler() http.HandlerFunc {
+	return func(w http.ResponseWriter, _ *http.Request) {
+		if ready {
+			w.Header().Set("Content-Type", "text/plain; charset=utf-8")
+			w.WriteHeader(http.StatusOK)
+			w.Write([]byte("ready"))
+		} else {
+			w.Header().Set("Content-Type", "text/plain; charset=utf-8")
+			w.WriteHeader(http.StatusServiceUnavailable)
+			w.Write([]byte("not ready"))
+		}
+	}
+}